% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_assess_estimates}
\alias{get_assess_estimates}
\alias{get_assess_estimates.MSE}
\alias{get_assess_estimates.list}
\title{Title}
\usage{
get_assess_estimates(x, ...)

\method{get_assess_estimates}{MSE}(x, model = "Model 1")

\method{get_assess_estimates}{list}(x)
}
\arguments{
\item{...}{}
}
\description{
Title
}
