% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_Biomass}
\alias{get_Biomass}
\alias{get_Biomass.Hist}
\alias{get_Biomass.MSE}
\alias{get_Biomass.list}
\title{Title}
\usage{
get_Biomass(x, ...)

\method{get_Biomass}{Hist}(x, name = "Name")

\method{get_Biomass}{MSE}(x, name = "Name")

\method{get_Biomass}{list}(x, name = "Name")

\method{get_Biomass}{MSE}(x, name = "Name")
}
\arguments{
\item{...}{}
}
\description{
Title
}
