% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_Landings}
\alias{get_Landings}
\alias{get_Landings.Hist}
\title{Title}
\usage{
get_Landings(x, ...)

\method{get_Landings}{Hist}(x, name = "Name")
}
\arguments{
\item{...}{}
}
\description{
Title
}
