% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_LifeHistory}
\alias{get_LifeHistory}
\alias{get_LifeHistory.Hist}
\alias{get_LifeHistory.list}
\alias{get_LifeHistory.MSE}
\title{Title}
\usage{
get_LifeHistory(x, model = "Model 1")

\method{get_LifeHistory}{Hist}(x, model = "Model 1")

\method{get_LifeHistory}{list}(x, model = "Model 1")

\method{get_LifeHistory}{MSE}(x, model = "Model 1")
}
\arguments{
\item{model}{}
}
\description{
Title
}
