% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_SSB}
\alias{get_SSB}
\alias{get_SSB.Hist}
\alias{get_SSB.MSE}
\alias{get_SSB.list}
\title{Title}
\usage{
get_SSB(x, name = "Name")

\method{get_SSB}{Hist}(x, name = "Name")

\method{get_SSB}{MSE}(x, name = "Name")

\method{get_SSB}{list}(x, name = "Name")
}
\arguments{
\item{name}{}
}
\description{
Title
}
