% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_ts}
\alias{get_ts}
\alias{valid_ts_variables}
\alias{match_ts_variable}
\alias{get_ts.Hist}
\alias{get_ts.multiHist}
\alias{get_ts.MSE}
\alias{get_ts.list}
\title{Create a data.frame with time-series information by simulation and year}
\usage{
get_ts(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)

valid_ts_variables()

match_ts_variable(variable = "Spawning Biomass", class = "Hist")

\method{get_ts}{Hist}(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)

\method{get_ts}{multiHist}(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)

\method{get_ts}{MSE}(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)

\method{get_ts}{list}(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)
}
\arguments{
\item{x}{An object of class \code{Hist}, \code{MSE}, or a list of \code{Hist} or \code{MSE} objects}

\item{variable}{A character string with a valid name for a time-series variable.
Use \code{valid_ts_variables()} for valid variable names.}

\item{model}{An optional name for the model.
If \code{x} is a list of objects, \code{model} will be taken from \code{names(x)}.
If \code{names(x)} is \code{NULL}, \code{model} will be given sequential numerical values (e.g., \verb{Model 1}, \verb{Model 2}, ...)}

\item{scale}{An optional function that scales the values in \code{Values}. See \code{lb2kg} and \code{kg2lb} for example.}
}
\description{
Create a data.frame with time-series information by simulation and year
}
