% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_years}
\alias{get_years}
\alias{get_years.MSE}
\alias{get_years.Hist}
\alias{get_years.multiHist}
\title{Create a data.frame with Historical and Projection years}
\usage{
get_years(x)

\method{get_years}{MSE}(x)

\method{get_years}{Hist}(x)

\method{get_years}{multiHist}(x)
}
\arguments{
\item{x}{An object of class `Hist`, `MSE`, `multiHist`, or `MMSE`}
}
\value{
A data.frame with years and period (Historical or Projection)
}
\description{
Create a data.frame with Historical and Projection years
}
