% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets_rds.R
\docType{data}
\name{Hist3}
\alias{Hist3}
\title{Example objects for testing}
\format{
An object of class \code{Hist} of length 1.
}
\usage{
Hist3
}
\description{
Will be deleted before production
}
\keyword{datasets}
