% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets_rds.R
\docType{data}
\name{MSE2}
\alias{MSE2}
\title{Example objects for testing}
\format{
An object of class \code{MSE} of length 1.
}
\usage{
MSE2
}
\description{
Will be deleted before production
}
\keyword{datasets}
