% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets_rds.R
\docType{data}
\name{MSEList}
\alias{MSEList}
\title{Example objects for testing}
\format{
An object of class \code{list} of length 4.
}
\usage{
MSEList
}
\description{
Will be deleted before production
}
\keyword{datasets}
