% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_Biomass}
\alias{plot_Biomass}
\alias{plot_Biomass.Hist}
\alias{plot_Biomass.list}
\alias{plot_Biomass.MSE}
\title{Title}
\usage{
plot_Biomass(
  x,
  title = "",
  xlab = "Year",
  ylab = "Biomass",
  quantiles = c(0.025, 0.975),
  scale = NULL,
  ...
)

plot_Biomass.Hist(x, ...)

plot_Biomass.list(x, ...)

plot_Biomass.MSE(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
