% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_Landings}
\alias{plot_Landings}
\alias{plot_Landings.Hist}
\alias{plot_Landings.list}
\alias{plot_Landings.MSE}
\title{Title}
\usage{
plot_Landings(
  x,
  title = "",
  xlab = "Year",
  ylab = "Landings",
  quantiles = c(0.025, 0.975),
  scale = NULL,
  ...
)

plot_Landings.Hist(x, ...)

plot_Landings.list(x, ...)

plot_Landings.MSE(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
