% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_Recruits}
\alias{plot_Recruits}
\alias{plot_Recruits.Hist}
\alias{plot_Recruits.list}
\alias{plot_Recruits.MSE}
\title{Title}
\usage{
plot_Recruits(
  x,
  title = "",
  xlab = "Year",
  ylab = "Recruits",
  quantiles = c(0.025, 0.975),
  scale = NULL,
  ...
)

plot_Recruits.Hist(x, ...)

plot_Recruits.list(x, ...)

plot_Recruits.MSE(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
