% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_Removals}
\alias{plot_Removals}
\alias{plot_Removals.Hist}
\alias{plot_Removals.list}
\alias{plot_Removals.MSE}
\title{Title}
\usage{
plot_Removals(
  x,
  title = "",
  xlab = "Year",
  ylab = "Removals",
  quantiles = c(0.025, 0.975),
  ...
)

plot_Removals.Hist(x, ...)

plot_Removals.list(x, ...)

plot_Removals.MSE(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
