% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_SSB}
\alias{plot_SSB}
\alias{plot_SSB.Hist}
\alias{plot_SSB.list}
\alias{plot_SSB.MSE}
\title{Title}
\usage{
plot_SSB(
  x,
  title = "",
  xlab = "Year",
  ylab = "Spawning Biomass",
  quantiles = c(0.025, 0.975),
  scale = NULL,
  ...
)

plot_SSB.Hist(x, ...)

plot_SSB.list(x, ...)

plot_SSB.MSE(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
