% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets_rds.R
\docType{data}
\name{TS_Variables}
\alias{TS_Variables}
\title{Valid time-series variables}
\format{
An object of class \code{data.frame} with 10 rows and 3 columns.
}
\usage{
TS_Variables
}
\description{
Valid time-series variables
}
\keyword{datasets}
