% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_at_age}
\alias{get_at_age}
\alias{get_at_age.Hist}
\alias{get_at_age.MSE}
\alias{get_at_age.list}
\alias{get_at_age.multiHist}
\title{Create a data.frame with at-age schedules by simulation and year}
\usage{
get_at_age(x, model = "Model 1")

\method{get_at_age}{Hist}(x, model = "Model 1")

\method{get_at_age}{MSE}(x, model = "Model 1")

\method{get_at_age}{list}(x, model = "Model 1")

\method{get_at_age}{multiHist}(x, model = "Model 1")
}
\arguments{
\item{x}{An object of class \code{Hist}, \code{MSE}, or a list of \code{Hist} or \code{MSE} objects}

\item{model}{An optional name for the model.
If \code{x} is a list of objects, \code{model} will be taken from \code{names(x)}.
If \code{names(x)} is \code{NULL}, \code{model} will be given sequential numerical values (e.g., \verb{Model 1}, \verb{Model 2}, ...)}
}
\description{
Create a data.frame with at-age schedules by simulation and year
}
