% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_Biomass}
\alias{get_Biomass}
\alias{get_SSB}
\alias{get_Landings}
\alias{get_Removals}
\alias{get_Recruits}
\title{Title}
\usage{
get_Biomass(x, model = "Model 1", ...)

get_SSB(x, model = "Model 1", ...)

get_Biomass(x, model = "Model 1", ...)

get_Landings(x, model = "Model 1", ...)

get_Removals(x, model = "Model 1", ...)

get_Recruits(x, model = "Model 1", ...)
}
\arguments{
\item{name}{}
}
\description{
Title
}
