% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{valid_ts_variables}
\alias{valid_ts_variables}
\alias{match_ts_variable}
\alias{get_ts}
\alias{get_ts.Hist}
\alias{get_ts.multiHist}
\alias{get_ts.MSE}
\alias{get_ts.list}
\title{Title}
\usage{
valid_ts_variables()

match_ts_variable(variable = "Spawning Biomass", class = "Hist")

get_ts(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)

\method{get_ts}{Hist}(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)

\method{get_ts}{multiHist}(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)

\method{get_ts}{MSE}(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)

\method{get_ts}{list}(x, variable = "Spawning Biomass", model = "Model 1", scale = NULL)
}
\arguments{
\item{x}{}

\item{model}{}
}
\description{
Title
}
