% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_ts}
\alias{get_ts}
\alias{get_ts.Hist}
\alias{valid_ts_variables}
\alias{match_ts_variable}
\alias{get_ts.MSE}
\alias{get_ts.list}
\title{Title}
\usage{
get_ts(x, variable = "Spawning Biomass", model = "Model 1")

\method{get_ts}{Hist}(x, variable = "Spawning Biomass", model = "Model 1")

valid_ts_variables()

match_ts_variable(variable = "Spawning Biomass", class = "Hist")

\method{get_ts}{MSE}(x, variable = "Spawning Biomass", model = "Model 1")

\method{get_ts}{list}(x, variable = "Spawning Biomass", model = "Model 1")
}
\arguments{
\item{x}{}

\item{model}{}
}
\description{
Title
}
